requirements.txt
Django==2.2.7
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
